fix: ensure time matrix embeddings are float type for dropout

+++ python/model_mhc.py
@@ -353,8 +353,8 @@ class TiSASRec(torch.nn.Module):
 
         time_matrices = torch.LongTensor(time_matrices).to(self.dev)
         time_matrices = torch.clamp(time_matrices, min=0, max=self.time_span)
-        time_matrix_K = self.time_matrix_K_emb(time_matrices)
-        time_matrix_V = self.time_matrix_V_emb(time_matrices)
+        time_matrix_K = self.time_matrix_K_emb(time_matrices).float()
+        time_matrix_V = self.time_matrix_V_emb(time_matrices).float()
         time_matrix_K = self.time_matrix_K_dropout(time_matrix_K)
         time_matrix_V = self.time_matrix_V_dropout(time_matrix_V)
